extract out CreateLinkState interactor
