+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/state/PaymentElementLoader.kt
@@ -1,7 +1,6 @@
 package com.stripe.android.paymentsheet.state
 
 import android.os.Parcelable
-import com.stripe.android.DefaultCardBrandFilter
 import com.stripe.android.common.analytics.experiment.LogLinkHoldbackExperiment
 import com.stripe.android.common.coroutines.runCatching
 import com.stripe.android.common.model.CommonConfiguration
@@ -14,29 +13,18 @@ import com.stripe.android.core.utils.FeatureFlags
 import com.stripe.android.core.utils.UserFacingLogger
 import com.stripe.android.googlepaylauncher.GooglePayEnvironment
 import com.stripe.android.googlepaylauncher.GooglePayRepository
-import com.stripe.android.link.LinkAppearance
-import com.stripe.android.link.LinkConfiguration
-import com.stripe.android.link.account.LinkStore
-import com.stripe.android.link.gate.LinkGate
-import com.stripe.android.link.model.AccountStatus
-import com.stripe.android.link.model.toLoginState
-import com.stripe.android.link.ui.inline.LinkSignupMode
 import com.stripe.android.lpmfoundations.luxe.LpmRepository
-import com.stripe.android.lpmfoundations.luxe.isSaveForFutureUseValueChangeable
 import com.stripe.android.lpmfoundations.paymentmethod.CustomerMetadata
 import com.stripe.android.lpmfoundations.paymentmethod.CustomerMetadata.Permissions.Companion.createForPaymentSheetCustomerSession
 import com.stripe.android.lpmfoundations.paymentmethod.CustomerMetadata.Permissions.Companion.createForPaymentSheetLegacyEphemeralKey
 import com.stripe.android.lpmfoundations.paymentmethod.IS_PAYMENT_METHOD_SET_AS_DEFAULT_ENABLED_DEFAULT_VALUE
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentMethodMetadata
 import com.stripe.android.lpmfoundations.paymentmethod.PaymentSheetCardBrandFilter
-import com.stripe.android.lpmfoundations.paymentmethod.toPaymentSheetSaveConsentBehavior
 import com.stripe.android.model.ElementsSession
-import com.stripe.android.model.ElementsSession.Flag.ELEMENTS_MOBILE_FORCE_SETUP_FUTURE_USE_BEHAVIOR_AND_NEW_MANDATE_TEXT
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.model.PaymentMethod
 import com.stripe.android.model.SetupIntent
 import com.stripe.android.model.StripeIntent
-import com.stripe.android.paymentelement.confirmation.utils.sellerBusinessName
 import com.stripe.android.payments.core.analytics.ErrorReporter
 import com.stripe.android.payments.financialconnections.GetFinancialConnectionsAvailability
 import com.stripe.android.paymentsheet.PaymentSheet
@@ -153,11 +141,8 @@ internal class DefaultPaymentElementLoader @Inject constructor(
     private val eventReporter: EventReporter,
     private val errorReporter: ErrorReporter,
     @IOContext private val workContext: CoroutineContext,
-    private val retrieveCustomerEmail: RetrieveCustomerEmail,
-    private val accountStatusProvider: LinkAccountStatusProvider,
+    private val createLinkState: CreateLinkState,
     private val logLinkHoldbackExperiment: LogLinkHoldbackExperiment,
-    private val linkStore: LinkStore,
-    private val linkGateFactory: LinkGate.Factory,
     private val externalPaymentMethodsRepository: ExternalPaymentMethodsRepository,
     private val userFacingLogger: UserFacingLogger,
     private val cvcRecollectionHandler: CvcRecollectionHandler,
@@ -204,9 +189,9 @@ internal class DefaultPaymentElementLoader @Inject constructor(
 
         val linkState = async {
             createLinkState(
-                configuration = configuration,
                 elementsSession = elementsSession,
-                customer = customerInfo,
+                configuration = configuration,
+                customer = customerInfo?.toCustomerInfo(),
                 initializationMode = initializationMode,
                 linkAppearance = configuration.linkAppearance
             )
@@ -500,180 +485,6 @@ internal class DefaultPaymentElementLoader @Inject constructor(
         }
     }
 
-    private suspend fun createLinkState(
-        elementsSession: ElementsSession,
-        configuration: CommonConfiguration,
-        customer: CustomerInfo?,
-        initializationMode: PaymentElementLoader.InitializationMode,
-        linkAppearance: LinkAppearance?
-    ): LinkState? {
-        val linkConfig =
-            createLinkConfiguration(
-                configuration = configuration,
-                customer = customer,
-                elementsSession = elementsSession,
-                initializationMode = initializationMode,
-                linkAppearance = linkAppearance
-            ) ?: return null
-        return loadLinkState(
-            configuration = configuration,
-            linkConfiguration = linkConfig,
-            elementsSession = elementsSession,
-            linkSignUpDisabled = elementsSession.disableLinkSignup,
-        )
-    }
-
-    private suspend fun loadLinkState(
-        configuration: CommonConfiguration,
-        linkConfiguration: LinkConfiguration,
-        elementsSession: ElementsSession,
-        linkSignUpDisabled: Boolean,
-    ): LinkState {
-        val accountStatus = accountStatusProvider(linkConfiguration)
-
-        val loginState = accountStatus.toLoginState()
-
-        val isSaveForFutureUseValueChangeable = isSaveForFutureUseValueChangeable(
-            code = PaymentMethod.Type.Card.code,
-            intent = elementsSession.stripeIntent,
-            paymentMethodSaveConsentBehavior = elementsSession.toPaymentSheetSaveConsentBehavior(),
-            hasCustomerConfiguration = configuration.customer != null,
-        )
-        val hasUsedLink = linkStore.hasUsedLink()
-        val signupToggleEnabled = elementsSession.linkSignUpOptInFeatureEnabled
-
-        val linkSignupMode = when {
-            // If signup toggle enabled, we show a future usage + link combined toggle
-            signupToggleEnabled && !linkConfiguration.customerInfo.email.isNullOrBlank() -> {
-                LinkSignupMode.InsteadOfSaveForFutureUse
-            }
-            // If inline signup is disabled or user has used Link, we don't show inline signup
-            linkSignUpDisabled || hasUsedLink -> null
-            // If inline signup and save for future use, we show it alongside save for future use
-            isSaveForFutureUseValueChangeable -> LinkSignupMode.AlongsideSaveForFutureUse
-            // If inline signup and save for future usage is not displayed, only show link signup
-            else -> LinkSignupMode.InsteadOfSaveForFutureUse
-        }
-
-        return LinkState(
-            configuration = linkConfiguration,
-            loginState = loginState,
-            signupMode = linkSignupMode.takeIf {
-                val validFundingSource = linkConfiguration.stripeIntent.linkFundingSources
-                    .contains(PaymentMethod.Type.Card.code)
-
-                val notLoggedIn = accountStatus == AccountStatus.SignedOut
-
-                validFundingSource && notLoggedIn
-            },
-        )
-    }
-
-    @Suppress("LongMethod")
-    private suspend fun createLinkConfiguration(
-        configuration: CommonConfiguration,
-        customer: CustomerInfo?,
-        elementsSession: ElementsSession,
-        initializationMode: PaymentElementLoader.InitializationMode,
-        linkAppearance: LinkAppearance?,
-    ): LinkConfiguration? {
-        if (!configuration.link.shouldDisplay || !elementsSession.isLinkEnabled) {
-            return null
-        }
-
-        val isCardBrandFilteringRequired =
-            elementsSession.linkPassthroughModeEnabled &&
-                configuration.cardBrandAcceptance != PaymentSheet.CardBrandAcceptance.All
-
-        val cardBrandFilter =
-            if (isCardBrandFilteringRequired) {
-                PaymentSheetCardBrandFilter(configuration.cardBrandAcceptance)
-            } else {
-                DefaultCardBrandFilter
-            }
-
-        val shippingDetails = configuration.shippingDetails
-
-        val customerPhone = if (shippingDetails?.isCheckboxSelected == true) {
-            shippingDetails.phoneNumber
-        } else {
-            configuration.defaultBillingDetails?.phone
-        }
-
-        val customerEmail = retrieveCustomerEmail(
-            configuration = configuration,
-            customer = customer?.toCustomerInfo()
-        )
-
-        val customerInfo = LinkConfiguration.CustomerInfo(
-            name = configuration.defaultBillingDetails?.name,
-            email = customerEmail,
-            phone = customerPhone,
-            billingCountryCode = configuration.defaultBillingDetails?.address?.country,
-        )
-
-        val cardBrandChoice = elementsSession.cardBrandChoice?.let { cardBrandChoice ->
-            LinkConfiguration.CardBrandChoice(
-                eligible = cardBrandChoice.eligible,
-                preferredNetworks = cardBrandChoice.preferredNetworks,
-            )
-        }
-
-        val linkConfiguration = LinkConfiguration(
-            stripeIntent = elementsSession.stripeIntent,
-            merchantName = configuration.merchantDisplayName,
-            sellerBusinessName = initializationMode.sellerBusinessName,
-            merchantCountryCode = elementsSession.merchantCountry,
-            merchantLogoUrl = elementsSession.merchantLogoUrl,
-            customerInfo = customerInfo,
-            shippingDetails = shippingDetails?.takeIf { it.isCheckboxSelected == true },
-            passthroughModeEnabled = elementsSession.linkPassthroughModeEnabled,
-            cardBrandChoice = cardBrandChoice,
-            cardBrandFilter = cardBrandFilter,
-            financialConnectionsAvailability = GetFinancialConnectionsAvailability(elementsSession = elementsSession),
-            flags = elementsSession.linkFlags,
-            useAttestationEndpointsForLink = elementsSession.useAttestationEndpointsForLink,
-            suppress2faModal = elementsSession.suppressLink2faModal,
-            disableRuxInFlowController = elementsSession.disableRuxInFlowController,
-            enableDisplayableDefaultValuesInEce = elementsSession.linkEnableDisplayableDefaultValuesInEce,
-            linkSignUpOptInFeatureEnabled = elementsSession.linkSignUpOptInFeatureEnabled,
-            linkSignUpOptInInitialValue = elementsSession.linkSignUpOptInInitialValue,
-            elementsSessionId = elementsSession.elementsSessionId,
-            initializationMode = initializationMode,
-            linkMode = elementsSession.linkSettings?.linkMode,
-            billingDetailsCollectionConfiguration = configuration.billingDetailsCollectionConfiguration,
-            defaultBillingDetails = configuration.defaultBillingDetails,
-            allowDefaultOptIn = elementsSession.allowLinkDefaultOptIn,
-            googlePlacesApiKey = configuration.googlePlacesApiKey,
-            collectMissingBillingDetailsForExistingPaymentMethods =
-            configuration.link.collectMissingBillingDetailsForExistingPaymentMethods,
-            allowUserEmailEdits = configuration.link.allowUserEmailEdits,
-            allowLogOut = configuration.link.allowLogOut,
-            skipWalletInFlowController = elementsSession.linkMobileSkipWalletInFlowController,
-            customerId = elementsSession.customer?.session?.customerId,
-            linkAppearance = linkAppearance,
-            saveConsentBehavior = elementsSession.toPaymentSheetSaveConsentBehavior(),
-            forceSetupFutureUseBehaviorAndNewMandate = elementsSession
-                .flags[ELEMENTS_MOBILE_FORCE_SETUP_FUTURE_USE_BEHAVIOR_AND_NEW_MANDATE_TEXT] == true,
-            linkSupportedPaymentMethodsOnboardingEnabled =
-            elementsSession.linkSettings?.linkSupportedPaymentMethodsOnboardingEnabled.orEmpty(),
-        )
-
-        val useWebLink = !linkGateFactory.create(linkConfiguration).useNativeLink
-
-        if (isCardBrandFilteringRequired && useWebLink) {
-            // CBF isn't currently supported in the web flow.
-            return null
-        }
-
-        val collectsExtraBillingDetails = configuration.billingDetailsCollectionConfiguration.collectsAnything
-        if (collectsExtraBillingDetails && useWebLink) {
-            // Extra billing details collection isn't currently supported in the web flow.
-            return null
-        }
-
-        return linkConfiguration
-    }
 
     private suspend fun isGooglePayReady(
         configuration: CommonConfiguration,
@@ -936,7 +747,7 @@ internal class DefaultPaymentElementLoader @Inject constructor(
         }
     }
 
-    private sealed interface CustomerInfo {
+    internal sealed interface CustomerInfo {
         val id: String
         val ephemeralKeySecret: String
